refactor: move Search into ItaliaTheme, add translations
