+++ locales/de/LC_MESSAGES/volto.po
@@ -47,6 +47,10 @@ msgstr ""
 msgid "Aspetto"
 msgstr ""
 
+#: components/ItaliaTheme/Search/Search
+msgid "Attenzione!"
+msgstr ""
+
 #: components/ItaliaTheme/View/Commons/PageHeader/PageHeaderBando
 msgid "Bando"
 msgstr ""
@@ -208,6 +212,10 @@ msgstr ""
 msgid "Mostra lo sfondo del blocco"
 msgstr ""
 
+#: components/ItaliaTheme/Search/Search
+msgid "Nessun risultato ottenuto"
+msgstr ""
+
 #: components/ItaliaTheme/Blocks/NewsHome/Sidebar
 msgid "News"
 msgstr ""
@@ -265,6 +273,14 @@ msgstr ""
 msgid "Scroll to top"
 msgstr ""
 
+#: components/ItaliaTheme/Search/Search
+msgid "Search results"
+msgstr ""
+
+#: components/ItaliaTheme/Search/Search
+msgid "Search site"
+msgstr ""
+
 #: components/ItaliaTheme/Blocks/EventSearch/Sidebar
 msgid "SearchEventsBlock"
 msgstr ""
@@ -296,6 +312,10 @@ msgstr ""
 #: components/ItaliaTheme/Search/SearchCTs
 #: components/ItaliaTheme/Search/SearchTopics
 msgid "Show all"
+msgstr ""
+
+#: components/ItaliaTheme/Search/Search
+msgid "Sono occorsi degli errori"
 msgstr ""
 
 #: components/ItaliaTheme/Footer/FooterNewsletterSubscribe
@@ -941,6 +961,10 @@ msgstr ""
 msgid "filters"
 msgstr ""
 
+#: components/ItaliaTheme/Search/Search
+msgid "filtersCollapse"
+msgstr ""
+
 #: components/ItaliaTheme/Blocks/EventSearch/Body
 msgid "find"
 msgstr ""
@@ -1054,6 +1078,10 @@ msgstr ""
 
 #: components/ItaliaTheme/View/PersonaView/PersonaView
 msgid "foto_attivita_politica"
+msgstr ""
+
+#: components/ItaliaTheme/Search/Search
+msgid "found_n_results"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Commons/CardCalendar
@@ -1268,10 +1296,12 @@ msgid "optionActiveContentButton"
 msgstr ""
 
 #: components/ItaliaTheme/Header/SearchModal
+#: components/ItaliaTheme/Search/Search
 msgid "optionActiveContentInfo"
 msgstr ""
 
 #: components/ItaliaTheme/Header/SearchModal
+#: components/ItaliaTheme/Search/Search
 msgid "optionActiveContentLabel"
 msgstr ""
 
@@ -1280,6 +1310,7 @@ msgid "optionDateEnd"
 msgstr ""
 
 #: components/ItaliaTheme/Header/SearchModal
+#: components/ItaliaTheme/Search/Search
 msgid "optionDateEndButton"
 msgstr ""
 
@@ -1292,10 +1323,12 @@ msgid "optionDateStart"
 msgstr ""
 
 #: components/ItaliaTheme/Header/SearchModal
+#: components/ItaliaTheme/Search/Search
 msgid "optionDateStartButton"
 msgstr ""
 
 #: components/ItaliaTheme/Header/SearchModal
+#: components/ItaliaTheme/Search/Search
 msgid "options"
 msgstr ""
 
@@ -1306,6 +1339,10 @@ msgstr ""
 #: components/ItaliaTheme/View/UOView/UOView
 #: components/ItaliaTheme/View/VenueView/VenueView
 msgid "orario_pubblico"
+msgstr ""
+
+#: components/ItaliaTheme/Search/Search
+msgid "order_by"
 msgstr ""
 
 #: components/ItaliaTheme/View/PersonaView/PersonaView
@@ -1413,6 +1450,7 @@ msgid "related_uo"
 msgstr ""
 
 #: components/ItaliaTheme/Header/SearchModal
+#: components/ItaliaTheme/Search/Search
 msgid "removeOption"
 msgstr ""
 
@@ -1577,6 +1615,14 @@ msgstr ""
 msgid "searchLabel"
 msgstr ""
 
+#: components/ItaliaTheme/Search/Search
+msgid "search_adv_filters"
+msgstr ""
+
+#: components/ItaliaTheme/Search/Search
+msgid "search_content_types"
+msgstr ""
+
 #: components/ItaliaTheme/Blocks/SearchSections/SideBar
 msgid "search_service_block_sections"
 msgstr ""
@@ -1637,7 +1683,12 @@ msgstr ""
 msgid "searchsectionswidget-title"
 msgstr ""
 
+#: components/ItaliaTheme/Search/Search
+msgid "section_undefined"
+msgstr ""
+
 #: components/ItaliaTheme/Header/SearchModal
+#: components/ItaliaTheme/Search/Search
 msgid "sections"
 msgstr ""
 
@@ -1728,6 +1779,18 @@ msgstr ""
 #: components/ItaliaTheme/SkipLinks/SkipLinks
 #: components/ItaliaTheme/SkipLinksC/SkipLinksC
 msgid "skiplink-navigation"
+msgstr ""
+
+#: components/ItaliaTheme/Search/Search
+msgid "sort_on_date"
+msgstr ""
+
+#: components/ItaliaTheme/Search/Search
+msgid "sort_on_relevance"
+msgstr ""
+
+#: components/ItaliaTheme/Search/Search
+msgid "sort_on_title"
 msgstr ""
 
 #: components/ItaliaTheme/View/PersonaView/PersonaView
@@ -1825,6 +1888,7 @@ msgid "to"
 msgstr ""
 
 #: components/ItaliaTheme/Header/SearchModal
+#: components/ItaliaTheme/Search/Search
 #: components/ItaliaTheme/View/Commons/PageHeader/PageHeaderTassonomiaArgomenti
 msgid "topics"
 msgstr ""
